Move table-related stuff out into their own modules.
